+++ tests/testthat/test-setFishing.R
@@ -58,6 +58,11 @@ test_that("validEffort works", {
     # A scrambled vector is put in the right order
     ies <- ie[c(2,3,1,4)]
     expect_identical(validEffortVector(ies, params), ie)
+    # NA's are replaced by default
+    ie[2] <- 0
+    iesn <- ie
+    iesn[2] <- NA
+    expect_identical(validEffortVector(iesn, params), ie)
     # A single number is converted into a constant vector
     ie[] <- 2
     expect_identical(validEffortVector(2, params), ie)